Add storage modules for conversation history and session management
